+++ sdk/resourcemanager/botservice/armbotservice/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 0.6.0 (2022-05-31)
+### Features Added
+
+- New const `EmailChannelAuthMethodGraph`
+- New const `EmailChannelAuthMethodPassword`
+- New function `*OutlookChannel.GetChannel() *Channel`
+- New function `PossibleEmailChannelAuthMethodValues() []EmailChannelAuthMethod`
+- New struct `CreateEmailSignInURLResponse`
+- New struct `CreateEmailSignInURLResponseProperties`
+- New struct `EmailClientCreateSignInURLOptions`
+- New struct `EmailClientCreateSignInURLResponse`
+- New struct `OutlookChannel`
+- New field `AuthMethod` in struct `EmailChannelProperties`
+- New field `MagicCode` in struct `EmailChannelProperties`
+
+
 ## 0.5.0 (2022-05-17)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/botservice/armbotservice` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 0.5.0, which contains breaking changes.